+++ code/src/ui/interaction/form/form-utils.ts
@@ -20,6 +20,7 @@ import {
   PropertyShapeOrGroup,
   QuickViewFields,
   QuickViewGroupings,
+  ShaclDefaultDateValueMap,
   SparqlResponseField,
   TYPE_KEY,
   VALUE_KEY
@@ -422,6 +423,7 @@ function initFormField(
   };
 }
 
+
 /**
  * Get the default value based on the inputs. If default value is given, this will be return, otherwise,
  * it depends on the field name.
@@ -480,6 +482,16 @@ export function getDefaultVal(
     }
     // Default value can be null, and should return false if null
     return !!defaultValue;
+  }
+
+  if (defaultValue === ShaclDefaultDateValueMap.TOMORROW) {
+    const tomorrow = new Date();
+    tomorrow.setDate(tomorrow.getDate() + 1);
+    return tomorrow.toISOString().split("T")[0];
+  } else if (defaultValue === ShaclDefaultDateValueMap.START_OF_YEAR) {
+    return `${new Date().getFullYear()}-01-01`;
+  } else if (defaultValue === ShaclDefaultDateValueMap.START_OF_MONTH) {
+    return `${new Date().getFullYear()}-${(new Date().getMonth() + 1).toString().padStart(2, '0')}-01`;
   }
   // Returns the default value if passed, or else, empty string
   return defaultValue ? defaultValue : "";